Cleanup: move `ElasticsearchSinkTestsBase` into `Stubs` subfolder.
