+++ test/Serilog.Sinks.Elasticsearch.Tests/BulkActionTests.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using FluentAssertions;
 using Serilog.Events;
 using Serilog.Parsing;
+using Serilog.Sinks.Elasticsearch.Tests.Stubs;
 using Xunit;
 
 namespace Serilog.Sinks.Elasticsearch.Tests
