+++ test/Serilog.Sinks.Elasticsearch.Tests/ExceptionAsJsonObjectFormatterTests.cs
@@ -9,6 +9,7 @@ using Serilog.Events;
 using Serilog.Formatting.Elasticsearch;
 using Serilog.Parsing;
 using Serilog.Sinks.Elasticsearch.Tests.Domain;
+using Serilog.Sinks.Elasticsearch.Tests.Stubs;
 
 namespace Serilog.Sinks.Elasticsearch.Tests
 {
