+++ test/Serilog.Sinks.Elasticsearch.Tests/IndexDeciderTests.cs
@@ -4,6 +4,7 @@ using FluentAssertions;
 using Xunit;
 using Serilog.Events;
 using Serilog.Parsing;
+using Serilog.Sinks.Elasticsearch.Tests.Stubs;
 
 namespace Serilog.Sinks.Elasticsearch.Tests
 {
