+++ test/Serilog.Sinks.Elasticsearch.Tests/Templating/Sendsv8TemplateTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using FluentAssertions;
+using Serilog.Sinks.Elasticsearch.Tests.Stubs;
 using Xunit;
 
 namespace Serilog.Sinks.Elasticsearch.Tests.Templating
